Safari-609.1.11.1/LayoutTests/fast/harness/fastmallocstatistics-object.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script type="text/javascript">
shouldBeDefined(window.internals);

var stats = window.internals.mallocStatistics();
shouldBeDefined(stats);
shouldBeGreaterThanOrEqual("stats.reservedVMBytes", "0");
shouldBeGreaterThanOrEqual("stats.committedVMBytes", "0");
shouldBeGreaterThanOrEqual("stats.freeListBytes", "0");
</script>
</head>
<body>
</body>
</html>
