Safari-609.1.11.1/LayoutTests/fast/harness/fastmallocstatistics-object-expected.txt

PASS [object Internals] is defined.
PASS [object MallocStatistics] is defined.
PASS stats.reservedVMBytes is >= 0
PASS stats.committedVMBytes is >= 0
PASS stats.freeListBytes is >= 0
